qa: resolve Psalm issues
Most were due to the refactor to support both v1 and v2 of laminas-httphandlerrunner.
When possible, I've added explicit suppressions in the Psalm configuration, as these can then be found and removed when we update implementations for v4.
