+++ src/SwooleRequestHandlerRunner.php
@@ -5,8 +5,11 @@ declare(strict_types=1);
 namespace Mezzio\Swoole;
 
 use Laminas\HttpHandlerRunner\RequestHandlerRunner;
+use Mezzio\Swoole\Exception\InvalidArgumentException;
 use Mezzio\Swoole\RequestHandlerRunner\RequestHandlerConstantsInterface;
 use Mezzio\Swoole\RequestHandlerRunner\RequestHandlerRunnerTrait;
+use Psr\EventDispatcher\EventDispatcherInterface;
+use Swoole\Http\Server as SwooleHttpServer;
 
 /**
  * Starts a Swoole web server that handles incoming requests.
@@ -18,4 +21,16 @@ use Mezzio\Swoole\RequestHandlerRunner\RequestHandlerRunnerTrait;
 class SwooleRequestHandlerRunner extends RequestHandlerRunner implements RequestHandlerConstantsInterface
 {
     use RequestHandlerRunnerTrait;
+
+    public function __construct(
+        SwooleHttpServer $httpServer,
+        EventDispatcherInterface $dispatcher
+    ) {
+        // The HTTP server should not yet be running
+        if ($httpServer->getMasterPid() > 0 || $httpServer->getManagerPid() > 0) {
+            throw new InvalidArgumentException('The Swoole server has already been started');
+        }
+        $this->httpServer = $httpServer;
+        $this->dispatcher = $dispatcher;
+    }
 }
